Pull upstream changes from girder/girder into our fork (#2)
* Allow clients to request S3 accelerated transfers

S3 buckets transfer acceleration is an AWS feature that speeds up data
transfer from a client to an S3 bucket using CloudFront.

* Add a scope to allow tokens to list jobs

Permit non-admin users to list jobs that they could have already
accessed by id.  For some reason, the "all" endpoints were marked admin
access, even though there are jobs that a non-admin is allowed to access
that are not their own (jobs marked public, for instance).

* Switch from pkg_resources to importlib

This was complicated largely due to our test infrastructure for faux
plugins.

* Fix using importlib to find npm packages

This was an issue introduced in #3692

* Plugins should show short descriptions.

When switching away from pkg_resources, some plugins showed long
descriptions instead of short descriptions.

* Add an option to log to a file in the fuse mount

Otherwise, the only way to debug something in fuse is to run in
foreground mode, which fundamentally changes some aspects of how fuse
runs.  Logging to a file will incur an overhead, but not the same
changes or as much as foreground mode.

* Retry s3 requests when fetching from the assetstore

We find that occasionally the s3 request gets interrupted with a
ChunkedEncodingError.  This is marked as an IOError, but doesn't have an
errno set on it.  Both retry, in case the problem is recoverable, and
set an errno so downstream handlers (such as the fuse.py library) break
less.

* Simplify code and add comments

* Prevent S3 assetstore from sending requests with ad in path

Ad blockers may block requests to AWS having /ad/ in their path.
This MR makes the S3 assetstore regenerate a UID for the file in that
case.
Tested with Ublock Origin and AdBlock Plus with default settings.

* Pin cheroot

As of cheroot 11.0.0, the cheroot server creates a non-daemon thread
that prevents pytest from shutting down in some instances.  Pin cheroot
until the upstream library resolves this (see
https://github.com/cherrypy/cheroot/issues/769).

* Remove unneeded license classifiers

Since adding the license= fields, these are no longer necessary and just
cause warnings during packaging.

* As per circleci docs, migrate deploy steps to run steps.

See
https://circleci.com/docs/guides/orchestrate/migrate-from-deploy-to-run/

* Add pyproject.toml

This reduces warnings regarding pep 517.

* Pin bcrypt until passlib is updated

bcrypt 5.x makes a change that explicitly warns about truncating long
passwords.  passlib tests for a bug of some previous version by testing
with a long password, which raises this exception.

We currently use passlib for bcrypt and for one-time-passwords (2fa).
For the use of bcrypt, the change is straightforward.  We can get rid of
our `_cryptContext` object and replace any uses of it (two in
model/user.py and one in the rest user.py).  This will resolve the
current issue with passlib, and won't affect current salts, etc.

The OTP requires a bit more of a refactor, but we can switch to pyotp.
I haven't investigated whether this can be done without any major
change.

* Switch from passlib to pyotp for OTP

* Use bcrypt directly

Stop using passlib for bcrypt.

There is a slight change to how we speed up tests by using "plaintext"
(this is mimicked from the passlib library, though much simpler).  If we
actually use bcrypt for everything, the tests take nearly twice as long.

* Remove an unused import

* Clean up a test that invoked passlib

I'm not sure why this passed before

* Add an index to the file collection to speed up s3 deletions

This changes the generalized collector of indices to not explicitly use
sets, as we support compound indices

* Pin a deep subdependency

Pin get-intrinsic, which is required transitively by other packages,
such as webpack.  Version 1.3.1 breaks our build.

---------

Co-authored-by: Paul Choisel <paul.choisel@kitware.com>
Co-authored-by: David Manthey <david.manthey@kitware.com>